update: move selectors to distinct file
